+++ lib/features/saved_search/delete/saved_search_delete_reducer.dart
@@ -1,5 +1,5 @@
-import 'package:pass_emploi_app/features/saved_search/delete/actions.dart';
-import 'package:pass_emploi_app/features/saved_search/delete/state.dart';
+import 'package:pass_emploi_app/features/saved_search/delete/saved_search_delete_actions.dart';
+import 'package:pass_emploi_app/features/saved_search/delete/saved_search_delete_state.dart';
 
 SavedSearchDeleteState savedSearchDeleteReducer(SavedSearchDeleteState current, dynamic action) {
   if (action is SavedSearchDeleteLoadingAction) return SavedSearchDeleteLoadingState();
