+++ lib/features/saved_search/delete/saved_search_delete_state.dart
