+++ lib/presentation/saved_search/saved_search_delete_view_model.dart
@@ -1,5 +1,5 @@
-import 'package:pass_emploi_app/features/saved_search/delete/actions.dart';
-import 'package:pass_emploi_app/features/saved_search/delete/state.dart';
+import 'package:pass_emploi_app/features/saved_search/delete/saved_search_delete_actions.dart';
+import 'package:pass_emploi_app/features/saved_search/delete/saved_search_delete_state.dart';
 import 'package:pass_emploi_app/redux/states/app_state.dart';
 import 'package:redux/redux.dart';
