+++ lib/redux/reducers/app_reducer.dart
@@ -1,5 +1,5 @@
-import 'package:pass_emploi_app/features/saved_search/delete/actions.dart';
-import 'package:pass_emploi_app/features/saved_search/delete/reducer.dart';
+import 'package:pass_emploi_app/features/saved_search/delete/saved_search_delete_actions.dart';
+import 'package:pass_emploi_app/features/saved_search/delete/saved_search_delete_reducer.dart';
 import 'package:pass_emploi_app/models/immersion.dart';
 import 'package:pass_emploi_app/models/immersion_details.dart';
 import 'package:pass_emploi_app/models/offre_emploi.dart';
